+++ babel.config.js
@@ -45,6 +45,7 @@ module.exports = {
             ,"@shopify/draggable": `./${debugPath}/plugins/@shopify/draggable`
             ,"prism-es6": `./${debugPath}/plugins/prism-es6/prism.js`
             ,"nouislider": `./${debugPath}/plugins/nouislider/dist/nouislider.mjs`
+            ,"vanilla-lazyload": `./${debugPath}/plugins/vanilla-lazyload/dist/esm/lazyload.js`
           }
         }
       ]